Added More statistics, small cleanup + Integrated GeoIp location API https://app.abstractapi.com/api/ip-geolocation/tester

+++ backend/src/main/java/net/estg/ei/backend/dao/AbstractDAOImpl.java
@@ -21,7 +21,7 @@ public abstract class AbstractDAOImpl<T> implements IAbstractDAO<T>
   @PersistenceContext
   private EntityManager entityManager;
 
-  public T findById(int id) {
+  public T findById(long id) {
     return entityManager.find(clazz, id);
   }
